No route URL update
Closes #9

The `go` method has been expanded so that it will take an object as well
as a string. When a string is supplied it behaves as it did before however
when an object is supplied then the behavour of the router can be
modified. The expected structure of the object is:

```
{
	href: '/some-url',
	route: 'url-only'
}
```

Where `href` is the URL which the title bar will be set to and route is
used to alter the behavior of the router, currently the only accepted
option is 'url-only' when it is supplied only the URL for the page is
updated otherwise the router will try and match that to a component and
update the display.

The behavour of the router can now also be modified from within the view
by using the `data-route` attribute as follows:

```
<a href="/some-url" data-route="url-only">Link text</a>
```

When using the `data-route` attribute only the following values are
accepted.

* `url-only` - Behaves the same as the `go` method, only updates the
  browsers URL.
* `none` - Will cause the router to ignore the click and let the browser
  handle it.