Emit an event on route change.
`ko.singlepage` now exposes a `element` property which is the element that
Knockout-SinglePage is bound to. The exposed element is also the source
of the events emitted by Knockout-SinglePage, the only event which is
currently emitted is `ko-sp-route-changed` which is emitted when the URL
is changed in such a way that it loads a new component. The data that
should be cared about when this event is emitted is contained within the
`detail` property of the event and is structured like the following:

```
{
	detail: {
		url: '/new/url',
		component: 'component-name'
	}
}
```

A set of utility methods are provided to enable easier binding to the
events emitted by Knockout-SinglePage, they are as follows:

* `on(event, callback)` - This will bind the callback method specified in
  as the second parameter to the event specified in the first parameter.
* `onRouteChanged(callback)` - This is a shorthand method that is the
  equivilent of calling `ko.singlepage.on('ko-sp-route-changed',
  callback)`.
* `off(event, callback)` - This will unbind the callback from the event.
* `offRouteChanged(callback)` - This is the equivilent of
  `ko.singlepage.off('ko-sp-route-changed', callback)`.

There is also a utility property exposed called `events` that includes the
strings defining the event names which is strucutred like so:

```
{
	routeChanged: 'ko-sp-route-changed'
}
```