Emit an event on route change.
Refactored the `init` method to alternatively accept an object as the
argument passed in. When pasing in the object events can be bound before
the first component is loaded which means that code can react when the
first component is loaded. An example of passing an object as an argument
would look like the following:

```
ko.singlePage.init({
	routes: [],
	element: document.getElementById('app'),
	on: {
		routeChanged: function(e) { console.log(e); }
	}
});
```

The existing behaviour where the routes and optionally the element can be
passed in as individual arguments remains.