Format URL.
Added the `href` binding to allow users to format the links given a route
and parameters. Updated the demos to use the new binding. Updated the
`format` method of the Route class so that it no longer throws exceptions
when things aren't right as that causes issues when using this
functionality in a binding.